src/requirements.txt
Pillow>=10.0.0  # For image manipulation 